added mic input device option. reworked prediction code, now more generic. added cl_lerp_smooth, cl_predict_extrapolate, cl_predict_timenudge cvars to allow tweaking player prediction/smoothness in a few different ways. cl_lerp_smooth's default changed to not smooth out live games in order to avoid unnecessary lag (was effectively set to 1, and would be 0 in vanilla).
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4471 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/server/sv_phys.c
@@ -2090,11 +2090,11 @@ void World_Physics_Frame(world_t *w)
 				}
 				else
 				{
-					int newt;
-					int delt;
+					float newt;
+					float delt;
 					newt = sv.time*1000;
 					delt = newt - svs.clients[i-1].msecs;
-					if (delt > 1000/77 || delt < -10)
+					if (delt > 1000/77.0 || delt < -10)
 					{
 						float ft = host_frametime;
 						host_client = &svs.clients[i-1];
